core (formats): Use signed timestamps.
Signed timestamps better support gapless playback of audio and are
required for video decode timestamps.

The following new standards are set out for format readers:

- The number of frames and duration of a Track always exclude delay and
  padding frames.
- The PTS, DTS, and duration of a Packet should be as written by the
  container. If the timing information must be synthesized by the reader,
  then negative PTS should be used for delay frames. The duration should
  exclude both delay and padding frames. Encoder delay and padding to be
  trimmed should be signalled via the trim_start and trim_end fields. The
  sum of duration and trim fields equal the duration decoded. For audio
  tracks, the DTS should equal the PTS.
- Seeking to delay or padding PTS should be supported.
- Decoders are responsible for trimming the output if gapless playback is
  enabled.

Timestamps and durations are now represented by newtype wrappers of i64
and u64, respectively. The wrappers do not implement arithmetic operations
that may panic or wrap. Rather, checked arithmetic operations are provided
instead. Since i64 is not capable of representing a full unsigned 64-bit
range (as supported by many media containers), these interfaces ensure
overflows are always considered by the developer.

The following guidelines are put forward for handling timestamp overflow:

- Timestamp overflow should be lazily handled only once that timestamp is
  reached during playback or requested in a seek.
- When timestamp overflow occurs, a warning should be logged.
- A FormatReader::next_packet implementation should return Ok(None) since
  the stream has effectively ended.
- A FormatReader::seek implementation should return an out-of-range seek
  error since the stream effectively ended before the overflow.
- A MetadataReader should ignore unrepresentable timestamps (e.g., ignore
  a chapter whose start time overflows).

The guiding principle for timestamp overflow is to support playing as much
of the media as possible before bailing out. Given the range of an i64,
this should never realistically happen in real-world media streams.

+++ symphonia-bundle-mp3/src/header.rs
@@ -14,7 +14,7 @@ use crate::common::*;
 pub const MPEG_HEADER_LEN: usize = 4;
 
 /// The maximum length in bytes of a MPEG audio frame including the header.
-pub const MAX_MPEG_FRAME_SIZE: u64 = 2881;
+pub const MAX_MPEG_FRAME_SIZE: usize = 2881;
 
 /// Bit-rate lookup table for MPEG version 1 layer 1.
 const BIT_RATES_MPEG1_L1: [u32; 15] = [
